editoast: rename models::projects to models::project
